refactor some files
